use input device manager for screenshare
