+++ packages/react-native-sdk/src/components/Call/CallControls/ScreenShareToggleButton.tsx
@@ -5,17 +5,7 @@ import { ScreenShare } from '../../../icons/ScreenShare';
 import { StopScreenShare } from '../../../icons/StopScreenShare';
 import { CallControlsButton } from './CallControlsButton';
 import { useTheme } from '../../../contexts/ThemeContext';
-import { useCall, useCallStateHooks } from '@stream-io/video-react-bindings';
-import { useIsIosScreenshareBroadcastStarted } from '../../../hooks/useIsIosScreenshareBroadcastStarted';
-import { hasScreenShare, SfuModels } from '@stream-io/video-client';
-import { usePrevious } from '../../../utils/hooks/usePrevious';
-
-// ios >= 14.0 or android - platform restrictions
-const CanDeviceScreenShare =
-  (Platform.OS === 'ios' &&
-    // @ts-ignore
-    Number.parseInt(Platform.Version.split('.')[0], 10) >= 14) ||
-  Platform.OS === 'android';
+import { useScreenShareButton } from '../../../hooks/useScreenShareButton';
 
 /**
  * The props for the Screen Share button in the Call Controls.
@@ -45,86 +35,18 @@ export const ScreenShareToggleButton = ({
     theme: { colors, screenShareToggleButton },
   } = useTheme();
 
-  const call = useCall();
-  const { useLocalParticipant, useCallSettings } = useCallStateHooks();
-  const callSettings = useCallSettings();
-  const isScreenSharingEnabledInCall = callSettings?.screensharing.enabled;
-  const isScreenSharingAccessRequestEnabled =
-    callSettings?.screensharing.access_request_enabled;
+  const screenCapturePickerViewiOSRef = useRef(null);
 
-  const onScreenShareStartedHandlerRef = useRef(onScreenShareStartedHandler);
-  onScreenShareStartedHandlerRef.current = onScreenShareStartedHandler;
-  const onScreenShareStoppedHandlerRef = useRef(onScreenShareStoppedHandler);
-  onScreenShareStoppedHandlerRef.current = onScreenShareStoppedHandler;
-
-  const iosScreenShareStartedFromSystem = useIsIosScreenshareBroadcastStarted();
-  const prevIosScreenShareStartedFromSystem = usePrevious(
-    iosScreenShareStartedFromSystem
+  const { onPress, hasPublishedScreenShare } = useScreenShareButton(
+    screenCapturePickerViewiOSRef,
+    onScreenShareStartedHandler,
+    onScreenShareStoppedHandler
   );
 
-  const localParticipant = useLocalParticipant();
-  const hasPublishedScreenShare =
-    localParticipant && hasScreenShare(localParticipant);
-
-  // listens to iOS screen share broadcast started event from the system
-  useEffect(() => {
-    const run = async () => {
-      if (Platform.OS !== 'ios') {
-        return;
-      }
-      if (
-        iosScreenShareStartedFromSystem &&
-        !prevIosScreenShareStartedFromSystem
-      ) {
-        onScreenShareStartedHandlerRef.current?.();
-        const media = await navigator.mediaDevices.getDisplayMedia();
-        await call?.publishScreenShareStream(media);
-      } else if (
-        !iosScreenShareStartedFromSystem &&
-        prevIosScreenShareStartedFromSystem
-      ) {
-        onScreenShareStoppedHandlerRef.current?.();
-        await call?.stopPublish(SfuModels.TrackType.SCREEN_SHARE);
-      }
-    };
-    run();
-  }, [
-    call,
-    iosScreenShareStartedFromSystem,
-    prevIosScreenShareStartedFromSystem,
-  ]);
-
-  const screenCapturePickerViewiOSRef = React.useRef(null);
-
-  const onPress = async () => {
-    if (!hasPublishedScreenShare) {
-      if (Platform.OS === 'ios') {
-        const reactTag = findNodeHandle(screenCapturePickerViewiOSRef.current);
-        await NativeModules.ScreenCapturePickerViewManager.show(reactTag);
-        // After this the iOS screen share broadcast started/stopped event will be triggered
-        // and the useEffect listener will handle the rest
-      } else {
-        try {
-          const media = await navigator.mediaDevices.getDisplayMedia();
-          onScreenShareStartedHandler?.();
-          await call?.publishScreenShareStream(media);
-        } catch (e) {
-          // ignored.. user didnt allow the screen share in the popup
-        }
-      }
-    } else if (hasPublishedScreenShare) {
-      onScreenShareStoppedHandler?.();
-      await call?.stopPublish(SfuModels.TrackType.SCREEN_SHARE);
-    }
-  };
-
-  if (!isScreenSharingEnabledInCall || !CanDeviceScreenShare) {
-    return null;
-  }
+  if (!onPress) return null;
 
   return (
     <CallControlsButton
-      disabled={!isScreenSharingAccessRequestEnabled}
       onPress={onPress}
       color={hasPublishedScreenShare ? colors.error : colors.static_white}
       style={{
